requirements.txt
fastapi==0.96.0
faiss==1.5.3
faiss_cpu==1.7.3
modal==0.49.2348
numpy==1.23.4
openai==0.27.0
pandas==1.5.2
python-dotenv==1.0.0
ratelimiter==1.2.0.post0
requests==2.28.1
scikit_learn==1.2.2
sentence_transformers==2.2.2
streamlit==1.23.1
tenacity==8.1.0
tqdm==4.64.1
